migrate com.sun.tools.javac API for ExtendedElementSupport and AbstractTreeScanner

+++ transpiler/src/main/java/org/jsweet/transpiler/SourcePosition.java
@@ -34,20 +34,15 @@ public final class SourcePosition {
 	/**
 	 * Creates a new source position from all indexes.
 	 * 
-	 * @param file
-	 *            the source file
-	 * @param sourceElement
-	 *            the source element if any
-	 * @param startLine
-	 *            the start line in the source file
-	 * @param startColumn
-	 *            the start column in the source file
-	 * @param endLine
-	 *            the end line in the source file
-	 * @param endColumn
-	 *            the end columb in the source file
+	 * @param file          the source file
+	 * @param sourceElement the source element if any
+	 * @param startLine     the start line in the source file
+	 * @param startColumn   the start column in the source file
+	 * @param endLine       the end line in the source file
+	 * @param endColumn     the end columb in the source file
 	 */
-	public SourcePosition(File file, Tree sourceElement, int startLine, int startColumn, int endLine, int endColumn) {
+	public SourcePosition(File file, Tree sourceElement, long startLine, long startColumn, long endLine,
+			long endColumn) {
 		super();
 		this.file = file;
 		this.startPosition = new Position(startLine, startColumn);
@@ -58,14 +53,10 @@ public final class SourcePosition {
 	/**
 	 * A simple constructor with no element and a single position (start = end).
 	 * 
-	 * @param file
-	 *            the source file
-	 * @param sourceElement
-	 *            the source element if any
-	 * @param line
-	 *            the position's line
-	 * @param column
-	 *            the position's column
+	 * @param file          the source file
+	 * @param sourceElement the source element if any
+	 * @param line          the position's line
+	 * @param column        the position's column
 	 */
 	public SourcePosition(File file, Tree sourceElement, int line, int column) {
 		this(file, sourceElement, new Position(line, column));
@@ -74,14 +65,10 @@ public final class SourcePosition {
 	/**
 	 * Creates a new source position from start and end positions.
 	 * 
-	 * @param file
-	 *            the source file
-	 * @param sourceElement
-	 *            the source element if any
-	 * @param startPosition
-	 *            the start position in the source file
-	 * @param endPosition
-	 *            the end position in the source file
+	 * @param file          the source file
+	 * @param sourceElement the source element if any
+	 * @param startPosition the start position in the source file
+	 * @param endPosition   the end position in the source file
 	 */
 	public SourcePosition(File file, Tree sourceElement, Position startPosition, Position endPosition) {
 		super();
@@ -92,15 +79,12 @@ public final class SourcePosition {
 	}
 
 	/**
-	 * Creates a new source position from a given position (will start and end
-	 * at the same position).
+	 * Creates a new source position from a given position (will start and end at
+	 * the same position).
 	 * 
-	 * @param file
-	 *            the source file
-	 * @param sourceElement
-	 *            the source element if any
-	 * @param position
-	 *            the position this source position with start and end at
+	 * @param file          the source file
+	 * @param sourceElement the source element if any
+	 * @param position      the position this source position with start and end at
 	 */
 	public SourcePosition(File file, Tree sourceElement, Position position) {
 		super();
@@ -138,28 +122,28 @@ public final class SourcePosition {
 	/**
 	 * The start line in the source file.
 	 */
-	public final int getStartLine() {
+	public final long getStartLine() {
 		return this.startPosition.getLine();
 	}
 
 	/**
 	 * The start column in the source file.
 	 */
-	public final int getStartColumn() {
+	public final long getStartColumn() {
 		return this.startPosition.getColumn();
 	}
 
 	/**
 	 * The end line in the source file.
 	 */
-	public final int getEndLine() {
+	public final long getEndLine() {
 		return this.endPosition.getLine();
 	}
 
 	/**
 	 * The end column in the source file.
 	 */
-	public final int getEndColumn() {
+	public final long getEndColumn() {
 		return this.endPosition.getColumn();
 	}
